Seperate watchdog into seperate header
